Workspace Html and css done
The workspace configuration (only the live sass compller) was set.
The object and their styles have been code.  They may need be revised, but it is a first try.
A bit of javascript was add to make the input-range slider functional and capable of  change the clock value when their values changed.